Ajout de tests pour la partie d'effacer des offres de stages si l'id est null ainsi que si l'id n'est pas trouvé

+++ src/main/java/fructose/service/OffreStageService.java
@@ -47,6 +47,12 @@ public class OffreStageService {
     }
 
     public void deleteOffreStage(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("ID ne peut pas être nul");
+        }
+        if (!offreStageRepository.existsById(id)) {
+            throw new IllegalArgumentException("OffreStage avec ID: " + id + " n'existe pas");
+        }
         offreStageRepository.deleteById(id);
     }
